+++ android/src/main/kotlin/com/airship/flutter/AirshipPlugin.kt
@@ -1,6 +1,5 @@
 package com.airship.flutter
 
-import android.app.Activity
 import android.graphics.Bitmap
 import android.webkit.WebView
 import android.app.NotificationManager
@@ -15,7 +14,6 @@ import com.urbanairship.json.JsonMap
 import com.urbanairship.json.JsonValue
 import com.urbanairship.channel.TagGroupsEditor
 import com.urbanairship.channel.AttributeEditor
-import com.urbanairship.iam.InAppMessageManager
 import com.urbanairship.messagecenter.MessageCenter
 import com.urbanairship.util.DateUtils
 import com.urbanairship.util.UAStringUtil
@@ -31,6 +29,8 @@ import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.BinaryMessenger
+import io.flutter.plugin.platform.PlatformViewRegistry
 
 private const val TAG_OPERATION_GROUP_NAME = "group"
 private const val TAG_OPERATION_TYPE = "operationType"
@@ -45,9 +45,9 @@ private const val ATTRIBUTE_MUTATION_VALUE = "value"
 private const val ATTRIBUTE_MUTATION_REMOVE = "remove"
 private const val ATTRIBUTE_MUTATION_SET = "set"
 
-class InboxMessageViewFactory(private val binding: FlutterPlugin.FlutterPluginBinding) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
+class InboxMessageViewFactory(private val binaryMessenger: BinaryMessenger) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
     override fun create(context: Context, viewId: Int, arguments: Any?): PlatformView {
-        val channel = MethodChannel(binding.flutterEngine.dartExecutor, "com.airship.flutter/InboxMessageView_$viewId")
+        val channel = MethodChannel(binaryMessenger, "com.airship.flutter/InboxMessageView_$viewId")
         val view = FlutterInboxMessageView(context, channel)
         channel.setMethodCallHandler(view)
         return view
@@ -122,21 +122,25 @@ class AirshipPlugin : MethodCallHandler, FlutterPlugin {
     companion object {
         @JvmStatic
         fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), "com.airship.flutter/airship")
-            channel.setMethodCallHandler(AirshipPlugin())
-            EventManager.shared.register(registrar)
+            var plugin = AirshipPlugin()
+            plugin.register(registrar.context().applicationContext, registrar.messenger(), registrar.platformViewRegistry())
         }
     }
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-        channel = MethodChannel(binding.flutterEngine.dartExecutor, "com.airship.flutter/airship")
-        binding.platformViewRegistry.registerViewFactory("com.airship.flutter/InboxMessageView", InboxMessageViewFactory(binding))
-        channel.setMethodCallHandler(this)
-        context = binding.applicationContext
+        register(binding.applicationContext, binding.binaryMessenger, binding.platformViewRegistry)
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
         channel.setMethodCallHandler(null)
+    }
+
+    private fun register(context: Context, binaryMessenger: BinaryMessenger, platformViewRegistry: PlatformViewRegistry) {
+        this.context = context
+        this.channel = MethodChannel(binaryMessenger, "com.airship.flutter/airship")
+        this.channel.setMethodCallHandler(this)
+        EventManager.shared.register(binaryMessenger)
+        platformViewRegistry.registerViewFactory("com.airship.flutter/InboxMessageView", InboxMessageViewFactory(binaryMessenger))
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {